+++ src/main/java/chess/domain/game/ChessGame.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import chess.domain.game.state.GameState;
 import chess.domain.game.state.StartedGame;
 import chess.domain.piece.Piece;
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Position;
 import chess.domain.piece.property.Color;
 
 public class ChessGame {
