+++ src/main/java/chess/domain/game/state/ChessBoard.java
@@ -8,10 +8,10 @@ import java.util.stream.Collectors;
 
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.position.Direction;
-import chess.domain.piece.position.File;
-import chess.domain.piece.position.Position;
-import chess.domain.piece.position.Rank;
+import chess.domain.game.state.position.Direction;
+import chess.domain.game.state.position.File;
+import chess.domain.game.state.position.Position;
+import chess.domain.game.state.position.Rank;
 import chess.domain.piece.property.Color;
 
 public class ChessBoard {
