+++ src/main/java/chess/domain/game/state/EndGame.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import chess.domain.piece.Piece;
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Position;
 import chess.domain.piece.property.Color;
 
 public class EndGame implements GameState {
