+++ src/main/java/chess/domain/game/state/RunningGame.java
@@ -3,7 +3,7 @@ package chess.domain.game.state;
 import java.util.Map;
 
 import chess.domain.piece.Piece;
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Position;
 import chess.domain.piece.property.Color;
 
 public class RunningGame extends StartedGame {
