+++ src/main/java/chess/domain/game/state/position/File.java
@@ -1,4 +1,4 @@
-package chess.domain.piece.position;
+package chess.domain.game.state.position;
 
 import java.util.Arrays;
