+++ src/main/java/chess/domain/game/state/position/Position.java
@@ -1,4 +1,4 @@
-package chess.domain.piece.position;
+package chess.domain.game.state.position;
 
 import java.util.HashMap;
 import java.util.Map;
