+++ src/main/java/chess/domain/piece/state/started/MovedPawn.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import chess.domain.game.state.ChessBoard;
-import chess.domain.piece.position.Direction;
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Direction;
+import chess.domain.game.state.position.Position;
 
 public class MovedPawn extends StartedPawn {
