+++ src/main/java/chess/domain/piece/state/started/StartedQueen.java
@@ -3,8 +3,8 @@ package chess.domain.piece.state.started;
 import java.util.List;
 
 import chess.domain.game.state.ChessBoard;
-import chess.domain.piece.position.Direction;
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Direction;
+import chess.domain.game.state.position.Position;
 import chess.domain.piece.state.PieceState;
 
 public class StartedQueen extends Started{
