+++ src/main/java/chess/dto/ChessBoardDto.java
@@ -3,7 +3,7 @@ package chess.dto;
 import java.util.HashMap;
 import java.util.Map;
 
-import chess.domain.piece.position.Position;
+import chess.domain.game.state.position.Position;
 import chess.domain.piece.Piece;
 
 public class ChessBoardDto {
